move CharsetVerifier to utils package
